Huge improvements to seeders to have actual testable database tables
Tables with actual id relationships which we can test on the website and
use to make in the future. Can use name-generated UUIDs to extend the
proper connections and be able to populate the tables with as many items
as we want. For now, we get two users and one record in every other
table.