Followed users now clickable, will take user to profile page of followed user which was clicked on
Added route, created index method in profilecontroller, got rid of an
unnecessary line in kitcontroller, made following page display links
that lead to each followed user's profile, made the profile list just
the user's name and the user's email for now. Frontend is just an ugly
test to show that the backend works.